fixed bug when total bank size is misestimated due to very large (e.g 2 Gbp axolotl chromosome) sequence

+++ gatb-core/src/gatb/bank/impl/BankFasta.cpp
@@ -766,7 +766,7 @@ void BankFasta::Iterator::estimate (u_int64_t& number, u_int64_t& totalSize, u_i
     {
         // linear extrapolation
         number    = (number    * _ref.getSize()) / actualPosition;
-        totalSize = (totalSize * _ref.getSize()) / actualPosition;
+        totalSize = totalSize * ( (float)_ref.getSize() / (float)actualPosition) ;
     }
 }
